+++ debug_toolbar/middleware.py
@@ -11,9 +11,7 @@ from django.utils.module_loading import import_string
 
 from debug_toolbar import settings as dt_settings
 from debug_toolbar.toolbar import DebugToolbar
-from debug_toolbar.utils import clear_stack_trace_caches
-
-_HTML_TYPES = ("text/html", "application/xhtml+xml")
+from debug_toolbar.utils import clear_stack_trace_caches, is_processable_html_response
 
 
 def show_toolbar(request):
@@ -102,13 +100,7 @@ class DebugToolbarMiddleware:
             response.headers[header] = value
 
         # Check for responses where the toolbar can't be inserted.
-        content_encoding = response.get("Content-Encoding", "")
-        content_type = response.get("Content-Type", "").split(";")[0]
-        if (
-            getattr(response, "streaming", False)
-            or content_encoding != ""
-            or content_type not in _HTML_TYPES
-        ):
+        if not is_processable_html_response(response):
             return response
 
         # Insert the toolbar in the response.